Decoupled lead login widget from main shortcode file
